pyproject.toml
[project]
name = "celestial-tts"
version = "0.1.0"
description = "A Text-to-Speech (TTS) multi-lingual and multi-provider REST API microservice for generating speech from text."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "flash-attn>=2.8.3",
    "hypercorn>=0.18.0",
    "pydantic-settings>=2.12.0",
    "qwen-tts>=0.0.4",
    "safetensors>=0.7.0",
    "soundfile>=0.13.1",
    "sqlmodel>=0.0.31",
    "tomli-w>=1.2.0",
    "torch>=2.10.0",
    "typer>=0.21.1",
    "uuid-utils>=0.14.0",
]
